feat : finish login and findPassword

+++ MiliMili/sending/views.py
@@ -11,7 +11,7 @@ from django.template import loader
 from user.models import User
 
 
-def send_email(token, email):
+def send_email(token, email, title):
     # 验证路由
     url = jwt.encode(token, SECRET_KEY, algorithm='HS256')  # 加密生成字符串
     if platform.system() == "Linux":
@@ -19,8 +19,13 @@ def send_email(token, email):
     else:
         url = os.path.join("http://127.0.0.1/api/sending/", url)
     data = {'url': url}
-    email_title = r"MiliMili邮箱激活"
-    email_body = loader.render_to_string('EmailContent-cloud.html', data)
+
+    if title == 'active':
+        email_title = r"MiliMili邮箱激活"
+        email_body = loader.render_to_string('EmailContent-register.html', data)
+    elif title == 'find':
+        email_title = r"MiliMili重设密码"
+        email_body = loader.render_to_string('EmailContent-find.html', data)
     try:
         msg = EmailMessage(email_title, email_body, settings.EMAIL_HOST_USER, [email])
         msg.content_subtype = 'html'
@@ -34,17 +39,27 @@ def active(request, url):
     # 获取token信息
     token = jwt.decode(url, SECRET_KEY, algorithms=['HS256'])
     user_id = token.get('user_id')
-    email = token.get('email')
-    # 激活用户 验证邮箱
-    user = User.objects.get(id=user_id)
-    user.isActive = True
-    user.email = email
-    user.save()
-    # 删除其他伪用户
-    username = user.username
-    user = User.objects.filter(username=username, isActive=False)
-    if user.exists():
-        user.delete()
     # 邮件的链接地址
     data = {'url': 'https://milimili.super2021.com'}
-    return render(request, 'EmailContent-call.html', data)
+
+    if 'email' in token.keys():
+        email = token.get('email')
+        # 激活用户 验证邮箱
+        user = User.objects.get(id=user_id)
+        user.isActive = True
+        user.email = email
+        user.save()
+        # 删除其他伪用户
+        username = user.username
+        user = User.objects.filter(username=username, isActive=False)
+        if user.exists():
+            user.delete()
+        return render(request, 'EmailContent-check-email.html', data)
+
+    if 'password' in token.keys():
+        password = token.get('password')
+        # 修改密码
+        user = User.objects.get(id=user_id)
+        user.password = password
+        user.save()
+        return render(request, 'EmailContent-check-find.html', data)
